% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynHier.R
\name{dynHier}
\alias{dynHier}
\title{dynHier}
\usage{
dynHier(x = NULL, ...)
}
\arguments{
\item{x}{if not NULL, one can specify a node created by {sdcTable::create_node()} or \code{cellKey::ck_create_node()}}

\item{...}{arguments (e.g \code{host}) that are passed through \code{\link[shiny]{runApp}} when
starting the shiny application}
}
\value{
The app can return a hierarchy object (either a \code{data.frame} or a tree-based object)
}
\description{
starts the \emph{shiny}-app to update/modify create a dynamic hierarchy that can be used for package \code{sdcTable}
and \code{cellKey}.
}
\examples{
\dontrun{
x <- dynHier(); print(x)

## start with a hierarchy-input
aDim <- sdcHier_create("Total", node_labs=letters[1:5])
aDim_mod <- dynHier(aDim); print(aDim_mod)
}
}
