% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixedHier.R
\name{fixedHier}
\alias{fixedHier}
\title{fixedHier}
\usage{
fixedHier(x, ...)
}
\arguments{
\item{x}{a character vector containing nested levels}

\item{...}{arguments (e.g \code{host}) that are passed through \code{\link[shiny]{runApp}} when
starting the shiny application}
}
\value{
The app can return a hierarchy object (either a \code{data.frame} or a tree-based object)
}
\description{
starts the \emph{shiny}-app to create a nested hierarchy starting from codes by defining positions
}
\examples{
\dontrun{
codes <- c("01","02","011","012")
res <- fixedHier(codes); print(res)
}
}
