% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hier_add.R
\name{hier_add}
\alias{hier_add}
\title{Add nodes to an existing hierarchy}
\usage{
hier_add(tree, node, leaves)
}
\arguments{
\item{tree}{a (nested) hierarchy created using \code{\link{hier_create}}
or modified using \code{\link{hier_add}}, \code{\link{hier_delete}}
or \code{\link{hier_rename}}.}

\item{node}{(character) an existing node in the input \code{h}}

\item{leaves}{names of the new nodes/levels that should be added}
}
\value{
a (nested) sdc hierarchy tree
}
\description{
This function allows to add nodes (levels)
to an existing nested hierarchy.
}
\examples{
## for examples, see hier_vignette()
}
\seealso{
hier_add hier_delete hier_rename hier_export
hier_convert hier_app hier_info
}
