% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hier_app.R
\name{hier_app}
\alias{hier_app}
\title{Create/Modify hierarchies interactively}
\usage{
hier_app(x, ...)
}
\arguments{
\item{x}{a character vector containing nested levels or
an object generated with \code{hier_create()}}

\item{...}{arguments (e.g \code{host}) that are passed
through \code{\link[shiny]{runApp}} when
starting the shiny application}
}
\value{
The app can return a hierarchy object (either
a \code{data.frame} or a tree-based object)
}
\description{
This function starts the \emph{shiny}-app to create a nested
hierarchy starting from codes by defining positions. Once
a hierarchy has been defined, it is possible to
modify and export it.
}
\examples{
\dontrun{
codes <- c("11", "12", "21", "22", "23", "31", "32")
res <- hier_app(codes); print(res)
}
}
