% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hier_codes.R
\name{hier_codes}
\alias{hier_codes}
\title{Default-Codes}
\usage{
hier_codes(tree)
}
\arguments{
\item{tree}{a (nested) hierarchy created using \code{\link[=hier_create]{hier_create()}}
or modified using \code{\link[=hier_add]{hier_add()}}, \code{\link[=hier_delete]{hier_delete()}} or \code{\link[=hier_rename]{hier_rename()}}.}
}
\value{
a named character vector with names being the node-names and the
values the standardized codes
}
\description{
\code{\link[=hier_codes]{hier_codes()}} returns the standardized codes for the nodes of a tree.
}
\examples{
h <- hier_create(root = "Total",  nodes = LETTERS[1:3])
h <- hier_add(h, root = "A", nodes = c("a1", "a5"))
hier_codes(h)
}
