% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hier_create.R
\name{hier_create}
\alias{hier_create}
\title{creates a nested hierarchy object}
\usage{
hier_create(tot_lab, node_labs = NULL)
}
\arguments{
\item{tot_lab}{(character) name of the overall total}

\item{node_labs}{(character) name of levels (nodes) in the hierarchy}
}
\value{
a (nested) sdc hierarchy tree
}
\description{
This functions allows to generate a hierarchical
data structure that can be used in other packages such
as \href{https://github.com/sdcTools/cellKey}{cellKey}
or \href{https://github.com/sdcTools/sdcTable}{sdcTable}.
}
\examples{
## for examples, see hier_vignette()
}
\seealso{
hier_add hier_delete hier_rename hier_export
hier_convert hier_app hier_info
}
