% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hier_delete.R
\name{hier_delete}
\alias{hier_delete}
\title{Delete nodes from an existing hierarchy}
\usage{
hier_delete(h, node_labs)
}
\arguments{
\item{h}{a (nested) hierarchy created using \code{\link{hier_create}} or modified
using \code{\link{hier_add}}, \code{\link{hier_delete}} or \code{\link{hier_rename}}.}

\item{node_labs}{names of the new nodes/levels that should be added}
}
\value{
a (nested) sdc hierarchy tree
}
\description{
This function allows to delete nodes (levels)
from an existing nested hierarchy.
}
\examples{
## for examples, see hier_vignette()
}
\seealso{
hier_add hier_delete hier_rename hier_export
hier_convert hier_app hier_info
}
