% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hier_display.R
\name{hier_display}
\alias{hier_display}
\title{Displays the hierarchy}
\usage{
hier_display(x, root = NULL)
}
\arguments{
\item{x}{a hierarchy object}

\item{root}{\code{NULL} if the entire tree should be printed or a name of
a node which is used as temporary root-node for printing}
}
\value{
\code{NULL}; the tree is printed to the prompt
}
\description{
Displays the hierarchy
}
\examples{
h <- hier_create(rootnode = "Total", leaves = LETTERS[1:2])
hier_display(h)
}
