% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hier_import.R
\name{hier_import}
\alias{hier_import}
\title{Imports a nested data structure}
\usage{
hier_import(inp, from = "json", tot_lab = NULL)
}
\arguments{
\item{inp}{an import from which should be converted. This could either
be a character-string (json) or a \code{data.frame}.}

\item{from}{(character) from which format should be imported.
Possible choices are:
\itemize{
\item \strong{"json"}: json-encoded string should be converted
\item \strong{"df"}: a \code{data.frame} in \code{@;level}-format will
be converted
\item \strong{"dt"}: a \code{data.table} in \code{@;level}-format will
be converted
\item \strong{"argus"}: an object exported using \code{\link{hier_convert}}
using \code{format = "argus"}
\item \strong{"code"}: an object exported using \code{\link{hier_convert}}
using \code{format = "code"}
\item \strong{"hrc"}: text-files in tau-argus hrc-format
\item \strong{"sdc"}: an object exported using \code{\link{hier_convert}}
using \code{format = "sdc"}
}}

\item{tot_lab}{optional name of overall total}
}
\value{
a (nested) hierarchy
}
\description{
This function creates a nested sdc hierarchy from various
input structures.
}
\examples{
## for examples, see hier_vignette()
}
