% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hier_nodenames.R
\name{hier_nodenames}
\alias{hier_nodenames}
\title{Extract name of nodes (levels)}
\usage{
hier_nodenames(tree, node_lab = NULL)
}
\arguments{
\item{tree}{a (nested) hierarchy created using \code{\link{hier_create}}
or modified using \code{\link{hier_add}}, \code{\link{hier_delete}}
or \code{\link{hier_rename}}.}

\item{node_lab}{(character) name of start node from which all lower
level-names should be returned}
}
\value{
a (nested) sdc hierarchy tree
}
\description{
This function allows to extract the all the names of the
nodes including all (sub)-nodes and leaves in the given
hierarchy.
}
\examples{
## for examples, see hier_vignette()
}
\seealso{
hier_add hier_delete hier_rename hier_export
hier_convert hier_app hier_info
}
