% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hier_rename.R
\name{hier_rename}
\alias{hier_rename}
\title{Rename nodes in an existing hierarchy}
\usage{
hier_rename(h, node_labs, node_labs_new)
}
\arguments{
\item{h}{a (nested) hierarchy created using \code{\link{hier_create}} or modified
using \code{\link{hier_add}}, \code{\link{hier_delete}} or \code{\link{hier_rename}}.}

\item{node_labs}{names of the new nodes/levels that should be added}

\item{node_labs_new}{(character) new names of nodes/levels that should be changed}
}
\value{
a (nested) sdc hierarchy tree
}
\description{
This function allows to rename one or more node(s) (levels)
in an existing nested hierarchy.
}
\examples{
## for examples, see hier_vignette()
}
\seealso{
hier_add hier_delete hier_rename hier_export
hier_convert hier_app hier_info
}
