% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hier_vignette.R
\name{hier_vignette}
\alias{hier_vignette}
\title{Show the package vignette}
\usage{
hier_vignette()
}
\value{
a browser windows/tab with showing the vignette
}
\description{
This function opens the introductionary package vignette
and opens it in a new browser tab/window.
}
\examples{
\dontrun{
hier_vignette()
}
}
