% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdcHier.R
\name{sdcHier}
\alias{sdcHier}
\title{sdcHier}
\usage{
sdcHier(x, ...)
}
\arguments{
\item{x}{a character vector containing nested levels or an object generated with \code{sdcHier_create()}}

\item{...}{arguments (e.g \code{host}) that are passed through \code{\link[shiny]{runApp}} when
starting the shiny application}
}
\value{
The app can return a hierarchy object (either a \code{data.frame} or a tree-based object)
}
\description{
starts the \emph{shiny}-app to create a nested hierarchy starting from codes by defining positions. Once
a hierarchy has been defined, it is possible to modify and export it.
}
\examples{
\dontrun{
codes <- c("11", "12", "21", "22", "23", "31", "32")
res <- sdcHier(codes); print(res)
}
}
