% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdcHier_add.R
\name{sdcHier_add}
\alias{sdcHier_add}
\title{sdcHier_add}
\usage{
sdcHier_add(h, refnode, node_labs)
}
\arguments{
\item{h}{a (nested) hierarchy created using \code{\link{sdcHier_create}} or modified
using \code{\link{sdcHier_add}}, \code{\link{sdcHier_delete}} or \code{\link{sdcHier_rename}}.}

\item{refnode}{(character) an existing node in the input \code{h}}

\item{node_labs}{names of the new nodes/levels that should be added}
}
\value{
a (nested) sdc hierarchy tree
}
\description{
add nodes/levels to a nested hierarchy
}
\examples{
## for examples, see sdcHier_vignette()
}
