% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdcHier_compute_fromList.R
\name{sdcHier_compute_fromList}
\alias{sdcHier_compute_fromList}
\title{sdcHier_compute_fromList}
\usage{
sdcHier_compute_fromList(dim, tot_lev, as_df = FALSE)
}
\arguments{
\item{dim}{a named list where each list-element containing codes of a hierarchical variables.}

\item{tot_lev}{a scalar characer specifying the name of the overall total}

\item{as_df}{(logical) if \code{FALSE}, a data.tree is returned, else a data.frame in \code{@;level}-format}
}
\value{
a hierarchical data structure depending on choice of argument \code{as_df}
}
\description{
compute a hierarchies from a list input. The name of each list-element refers to a (sub)-total (node) in the hierarchy
while the values of the each list-elements to the leafs contributing to this (sub)total (node)
}
\examples{
## nuts: digits 1-2 (nuts1), digit 3 (nut2), digits 4-5 (nuts3)
## the order of the list-elements is not important
## names of input-list correspond to sub-total names; its values to the contributing levels
geo_ll <- list()
geo_ll[["Total"]] <- c("01","02","03","10")
geo_ll[["010"]] <- c("01051","01053","01054","01055","01056","01057",
  "01058","01059","01060","01061","01062")
geo_ll[["031"]] <- c("03151","03152","03153","03154","03155","03156",
  "03157","03158")
geo_ll[["032"]] <- c("03251","03252","03254","03255","03256","03257")
geo_ll[["033"]] <- c("03351","03352","03353","03354","03355","03356",
  "03357","03358","03359","03360","03361")
geo_ll[["034"]] <- c("03451","03452","03453","03454","03455","03456")
geo_ll[["01"]] <- c("010")
geo_ll[["02"]] <- c("020")
geo_ll[["020"]] <- c("02000")
geo_ll[["03"]] <- c("031","032","033","034")
geo_ll[["10"]] <- c("101")
geo_ll[["101"]] <- c("10155")
d <- sdcHier_compute_fromList(dim=geo_ll, tot_lev="Total", as_df=FALSE); d
}
\seealso{
sdcHier_compute
}
