% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdcHier_convert.R
\name{sdcHier_convert}
\alias{sdcHier_convert}
\title{sdcHier_convert}
\usage{
sdcHier_convert(h, format = "data.frame")
}
\arguments{
\item{h}{a (nested) hierarchy created using \code{\link[=sdcHier_create]{sdcHier_create()}} or modified using \code{\link[=sdcHier_add]{sdcHier_add()}}, \code{\link[=sdcHier_delete]{sdcHier_delete()}} or
\code{\link[=sdcHier_rename]{sdcHier_rename()}}.}

\item{format}{(character) specifying the export format. possible choices are:
\itemize{
\item \strong{"data.frame"}: a \code{data.frame} with two columns. The first columns contains a string containing as many \code{@} as the level
of the node in the string (e.g \code{@} corresponds to the overall total while \code{@} would be all codes contributing to the total.
The second column contains the names of the levels.
\item \strong{"json"}: json format suitable as input for shiny Tree
}}
}
\value{
a (nested) sdc hierarchy tree
}
\description{
add nodes/levels to a nested hierarchy
}
\examples{
## for examples, see ?sdcHier_create
}
