% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdcHier_create.R
\name{sdcHier_create}
\alias{sdcHier_create}
\title{sdcHier_create}
\usage{
sdcHier_create(tot_lab, node_labs = NULL)
}
\arguments{
\item{tot_lab}{(character) name of the overall total}

\item{node_labs}{(character) name of levels (nodes) in the hierarchy}
}
\value{
a (nested) sdc hierarchy tree
}
\description{
generate and modify hierarchical data structure to be used in other packages such
as \href{https://github.com/sdcTools/cellKey}{cellKey}
or \href{https://github.com/sdcTools/sdcTable}{sdcTable}.
}
\examples{
## for examples, see sdcHier_vignette()
}
