% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdcHier_delete.R
\name{sdcHier_delete}
\alias{sdcHier_delete}
\title{sdcHier_delete}
\usage{
sdcHier_delete(h, node_labs)
}
\arguments{
\item{h}{a (nested) hierarchy created using \code{\link{sdcHier_create}} or modified
using \code{\link{sdcHier_add}}, \code{\link{sdcHier_delete}} or \code{\link{sdcHier_rename}}.}

\item{node_labs}{names of the new nodes/levels that should be added}
}
\value{
a (nested) sdc hierarchy tree
}
\description{
add nodes/levels from a nested hierarchy
}
\examples{
## for examples, see sdcHier_vignette()
}
