% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdcHier_import.R
\name{sdcHier_import}
\alias{sdcHier_import}
\title{sdcHier_import}
\usage{
sdcHier_import(inp, from = "json", tot_lab = NULL)
}
\arguments{
\item{inp}{an import from which should be converted. This could either
be a character-string (json).}

\item{from}{(character) from which format should be imported.
Possible choices are:
\itemize{
\item \emph{"json"}
}}

\item{tot_lab}{TODO}
}
\value{
a (nested) hierarchy suitable as input
for \code{sdcHier_xyz()}-functions
}
\description{
returns a tree from various input sources
}
\examples{
## not yet
}
