% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdcHier_info.R
\name{sdcHier_info}
\alias{sdcHier_info}
\title{sdcHier_info}
\usage{
sdcHier_info(h, node_labs = NULL)
}
\arguments{
\item{h}{a (nested) hierarchy created using \code{\link[=sdcHier_create]{sdcHier_create()}} or modified using \code{\link[=sdcHier_add]{sdcHier_add()}}, \code{\link[=sdcHier_delete]{sdcHier_delete()}} or
\code{\link[=sdcHier_rename]{sdcHier_rename()}}.}

\item{node_labs}{...}
}
\value{
\code{list}
}
\description{
get information about all or specific nodes in a nested hierarchy
}
\examples{
## for examples, see ?sdcHier_create
}
