% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdcHier_nodenames.R
\name{sdcHier_nodenames}
\alias{sdcHier_nodenames}
\title{sdcHier_nodenames}
\usage{
sdcHier_nodenames(h, node_lab = NULL)
}
\arguments{
\item{h}{a (nested) hierarchy created using \code{\link[=sdcHier_create]{sdcHier_create()}} or modified using \code{\link[=sdcHier_add]{sdcHier_add()}}, \code{\link[=sdcHier_delete]{sdcHier_delete()}} or
\code{\link[=sdcHier_rename]{sdcHier_rename()}}.}

\item{node_lab}{(character) name of start node from which all lower level-names should be returned}
}
\value{
a (nested) sdc hierarchy tree
}
\description{
get the names of the nodes and all (sub)-nodes and leaves of a given nested hierarchy
}
\examples{
## for examples, see ?sdcHier_create
}
