% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdcHier_rename.R
\name{sdcHier_rename}
\alias{sdcHier_rename}
\title{sdcHier_rename}
\usage{
sdcHier_rename(h, node_labs, node_labs_new)
}
\arguments{
\item{h}{a (nested) hierarchy created using \code{\link{sdcHier_create}} or modified
using \code{\link{sdcHier_add}}, \code{\link{sdcHier_delete}} or \code{\link{sdcHier_rename}}.}

\item{node_labs}{names of the new nodes/levels that should be added}

\item{node_labs_new}{(character) new names of nodes/levels that should be changed}
}
\value{
a (nested) sdc hierarchy tree
}
\description{
rename nodes/levels in a nested hierarchy
}
\examples{
## for examples, see ?sdcHier_create
}
