% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdcHier_vignette.R
\name{sdcHier_vignette}
\alias{sdcHier_vignette}
\title{sdcHier_vignette}
\usage{
sdcHier_vignette()
}
\value{
a browser windows/tab with showing the vignette
}
\description{
starts the package vignette that gets you started with the package
}
\examples{
\dontrun{
sdcHier_vignette()
}
}
