% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hier_create.R
\name{hier_create}
\alias{hier_create}
\title{Create a hierarchy}
\usage{
hier_create(root = "Total", nodes = NULL)
}
\arguments{
\item{root}{(character) name of the overall total}

\item{nodes}{(character) name of leaves (nodes) in the hierarchy}
}
\value{
a (nested) sdc hierarchy tree
}
\description{
This functions allows to generate a hierarchical
data structure that can be used in other packages such
as \href{https://github.com/sdcTools/cellKey}{\code{cellKey}}
or \href{https://github.com/sdcTools/sdcTable}{\code{sdcTable}}.
}
\examples{
# without nodes
h <- hier_create(root = "tot")
hier_display(h)

# with nodes
h <- hier_create(root = "tot", nodes = LETTERS[1:5])
hier_display(h)
}
\seealso{
hier_add hier_delete hier_rename hier_export
hier_convert hier_app hier_info
}
