% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hier_rename.R
\name{hier_rename}
\alias{hier_rename}
\title{Rename nodes in an existing hierarchy}
\usage{
hier_rename(tree, leaves)
}
\arguments{
\item{tree}{a (nested) hierarchy created using \code{\link{hier_create}}
or modified using \code{\link{hier_add}}, \code{\link{hier_delete}}
or \code{\link{hier_rename}}.}

\item{leaves}{(character) new names of nodes/levels that should be changed as
a named vector: names refer to old, existing names, the values to the
new labels}
}
\value{
a (nested) sdc hierarchy tree
}
\description{
This function allows to rename one or more node(s) (levels)
in an existing nested hierarchy.
}
\examples{
## for examples, see hier_vignette()
}
\seealso{
hier_add hier_delete hier_rename hier_export
hier_convert hier_app hier_info
}
