% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdcHier_convert.R
\name{sdcHier_convert}
\alias{sdcHier_convert}
\title{sdcHier_convert}
\usage{
sdcHier_convert(h, format = "data.frame", path = NULL,
  verbose = FALSE)
}
\arguments{
\item{h}{a (nested) hierarchy created using \code{\link{sdcHier_create}} or modified
using \code{\link{sdcHier_add}}, \code{\link{sdcHier_delete}} or \code{\link{sdcHier_rename}}.}

\item{format}{(character) specifying the export format. possible choices are:
\itemize{
\item \strong{"data.frame"}: a \code{data.frame} with two columns. The first columns contains a string
containing as many \code{@} as the level of the node in the string (e.g \code{@} corresponds to the overall
total while \code{@} would be all codes contributing to the total. The second column contains the names
of the levels.
\item \strong{"argus"}: used to create hrc-files suitable for tau-argus
\item \strong{"json"}: json format suitable as input for shiny Tree
\item \strong{"code"}: code required to generate the hierarchy
}}

\item{path}{if not \code{NULL}, the output is written to the file specified}

\item{verbose}{(logical) if true, the result of the conversion will not only
be (invisibly) returned but also printed in the prompt}
}
\value{
a (nested) sdc hierarchy tree
}
\description{
convert nested hierarchies into other data structures
}
\examples{
## for examples, see ?sdcHier_create
}
