% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdcHier_import.R
\name{sdcHier_import}
\alias{sdcHier_import}
\title{sdcHier_import}
\usage{
sdcHier_import(inp, from = "json", tot_lab = NULL)
}
\arguments{
\item{inp}{an import from which should be converted. This could either
be a character-string (json) or a \code{data.frame}.}

\item{from}{(character) from which format should be imported.
Possible choices are:
\itemize{
\item \strong{"json"}: json-encoded string should be converted
\item \strong{"df"}: a \code{data.frame} in \code{@;level}-format will be converted
\item \strong{"argus"}: an object exported using \code{\link{sdcHier_convert}} using \code{format = "argus"}
\item \strong{"code"}: an object exported using \code{\link{sdcHier_convert}} using \code{format = "code"}
\item \strong{"hrc"}: text-files in tau-argus hrc-format
\item \strong{"sdc"}: an object exported using \code{\link{sdcHier_convert}} using \code{format = "sdc"}
}}

\item{tot_lab}{optional name of overall total}
}
\value{
a (nested) hierarchy
}
\description{
returns a tree from various input sources
}
\examples{
## for examples, see sdcHier_vignette()
}
