% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdcHier_info.R
\name{sdcHier_info}
\alias{sdcHier_info}
\title{sdcHier_info}
\usage{
sdcHier_info(h, node_labs = NULL)
}
\arguments{
\item{h}{a (nested) hierarchy created using \code{\link{sdcHier_create}} or modified
using \code{\link{sdcHier_add}}, \code{\link{sdcHier_delete}} or \code{\link{sdcHier_rename}}.}

\item{node_labs}{names of the new nodes/levels that should be added}
}
\value{
a \code{list} with information about the required nodes. If \code{node_labs} is \code{NULL},
then the information is computed for all nodes.
}
\description{
get information about all or specific nodes in a nested hierarchy
}
\examples{
## for examples, see sdcHier_vignette()
}
