% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdcHier_info.R
\name{sdcHier_info}
\alias{sdcHier_info}
\title{sdcHier_info}
\usage{
sdcHier_info(h, node_labs = NULL)
}
\arguments{
\item{h}{a (nested) hierarchy created using \code{\link{sdcHier_create}} or modified
using \code{\link{sdcHier_add}}, \code{\link{sdcHier_delete}} or \code{\link{sdcHier_rename}}.}

\item{node_labs}{(character) name(s) of nodes for which information should be extracted.}
}
\value{
a \code{list} with information about the required nodes. If \code{node_labs} is
\code{NULL} (the default), the information is computed for all available nodes of
the hierarchy. The following properties are computed:
\itemize{
\item{exists: }{(logical) does the node exist}
\item{name: }{(character) node name}
\item{is_rootnode: }{(logical) is the node the overall root of the tree?}
\item{level: }{(numeric) what is the level of the nod}
\item{is_leaf: }{(logical) is the node a leaf?}
\item{siblings: }{(character) what are siblings of this node?}
\item{contributing_codes: }{(character) which codes are contributing to this node? If none (it is a leaf), \code{NA} is returned}
\item{children: }{(character) the names of the children of the node. If it has none (it is a leaf), \code{NA} is returned}
\item{is_bogus: }{(logical) is it a bogus code (i. e the only children of a leaf?)}
}
}
\description{
get information about all or specific nodes in a nested hierarchy
}
\examples{
## for examples, see sdcHier_vignette()
}
